+++ src/Collection/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Method `CollectionInterface::join()` to join items with a glue string and optional final glue
+
 ## [1.1.0] - 2025-11-21
 
 ### Changed